Refactored friendship model
Using denormalised/duplicate rows in “friends” table:
Each friendship relation will have fromId for both “sides” of the relationship.
Introduced a “requester” boolean to note the original “direction” of the request,
and show proposals differently in the UI.

This was necessary to keep DynamoDB scalable. Alternative would’ve been a secondary
global index on the “toId” range key, but that effectively doubles the provisioned throughput
cost for this table, and causes multiple requests for gathering friendships.

Also replaced uses of scan() with query(), which doesn’t need to map/reduce the entire table
to get results.