Fixed codepush
Somehow the ES6 decorator use (or maybe the wrapped render() call)
causes the camera not to work on the simulator.

I can't see how the @codePush decorator would change the behaviour of the camera,
but using it as a wrapped component does the trick.

For reference, the decorator is defined in https://github.com/Microsoft/react-native-code-push/blob/master/CodePush.js#L453