Reworked and documented fixtures
I prefer to only have the config.js file "dirty" on local,
rather than local changes splattered throughout the *.js files :)
The downside is that we're loading all fixtures even when they're not used.
There's little processing overhead here, and if it becomes a problem we'll make them function exports.