WIP Photo viewer
I figured that we could extract the relevant bits from react-native-photo-browser
if we felt it was getting too much of a burden. I don't link how restricted it is,
but am happy to work off a fork for a bit and see if the author accepts PRs.
Particularly around rendering a different photo view (with "delete" or "unfriend" buttons).
it could get a bit difficult to stick to the upstream library.

Regarding the navigator integration with photo-browser, it's a bit ganky:
It re-parses store.photos on detail view, and doesn't update the route
on photo change. But I think using TabNavigator just because it has
swipe abilities is the wrong answer here, react-navigator isn't set up
to have dynamic states like that AFAICT. We'll likely need a way
to deep link into photos or friends, but could do that through
one navigation screen with an optional "uuid" param, no need for a
separate navigator.

So the <ListView> usage in photo-browser is generally the right direction,
with horizontal scrolling, relative page offset calcs,
 and all the nice lazy loading that you get in this component.
We could replicate that in custom code, or look at something like https://github.com/leecade/react-native-swiper.

The lack of pinch to zoom in photo-browser isn't great, but there's pull requests adding that feature already.

Overall, I don't have a lot of confidence in the library, but it's a small amount of code
that I'd be happy rewriting in custom code if we really feel like it's going nowhere.